% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_snowflake_tat_data.R
\name{import_snowflake_tat_data}
\alias{import_snowflake_tat_data}
\title{Snowflake TAT Data Import Function}
\usage{
import_snowflake_tat_data(files)
}
\arguments{
\item{files}{A list of files that were exported from Snowflake.}
}
\description{
This function allows you import data from Snowflake.
}
