% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplr.R
\name{mplr}
\alias{mplr}
\title{Multiple Period Logistic Regression Function}
\usage{
mplr(data, output.file)
}
\arguments{
\item{data}{Data set that is used in the model estimation. Data set is obtained by using the data_cleaner() function.}

\item{output.file}{Name of a html file that the model result table is printed to. The file extension for this should be ".html"}
}
\description{
This function allows you to estimate multiple period logistic regression model, a mulitple period logistic regression model with backwards stepwise selection, and print out results.
}
