% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_regression.R
\name{survival_regression}
\alias{survival_regression}
\title{Survival Regression Function}
\usage{
survival_regression(data)
}
\arguments{
\item{data}{Data set to be used in model fitting. Data set is obtained from the data_cleaner() function.}
}
\description{
This function allows you to fit survival regression, survival regression with backwards stepwise selectoin, and print results.
}
